+++ TO2/Parser/ScriptParser.StringInterpolation.cs
@@ -21,8 +21,13 @@ public static class TO2ParserStringInterpolation {
         Tag("\\}").To('}')
     );
 
+    private static readonly Parser<string> AlignOrFormat =
+        Recognize(Seq(
+            Opt(Char(',').Then(Opt(Char('-')).Then(Digits1))),
+            Opt(Char(':').Then(CharsExcept1("\\\"\r\n{}", "align or format")))));
+
     public static IResult<Expression> StringInterpolationContentImpl(IInput input) => Many0(Alt<StringInterpolationPart>(
             Many1(ExtendedEscapedStringChar).Map(chars => new StringInterpolationPart.StringPart(new string(chars.ToArray()))),
-            Expression.Between(Char('{'), Char('}')).Map(expr => new StringInterpolationPart.ExpressionPart(expr))
-            )).Map((parts, start, end) => new StringInterpolation(parts, end, start))(input);
+            Seq(Expression, Opt(AlignOrFormat)).Between(Char('{'), Char('}')).Map(expr => new StringInterpolationPart.ExpressionPart(expr.Item1, expr.Item2.GetOrElse("")))
+        )).Map((parts, start, end) => new StringInterpolation(parts, end, start))(input);
 }